chore: refactor configuration management and logging client setup for the server
